requirements.txt
python-telegram-bot
PyYAML~=6.0
